+++ templates/serviceProviders-tpl.php
@@ -94,9 +94,6 @@ $lastDays = $this->data['lastDays'];
         }
     }
 </script>
-</head>
-
-<body>
 <div class="timeRange">
     <h4><?php echo $this->t('{proxystatistics:Proxystatistics:templates_time_range}'); ?></h4>
     <form id="dateSelector" method="post">
@@ -136,4 +133,3 @@ $lastDays = $this->data['lastDays'];
         <div id="spsTable" class="table"></div>
     </div>
 </div>
-</body>