+++ templates/summary-tpl.php
@@ -114,9 +114,6 @@ $lastDays = $this->data['lastDays'];
     }
 
 </script>
-</head>
-
-<body>
 <div class="timeRange">
     <h4><?php echo $this->t('{proxystatistics:Proxystatistics:templates_time_range}'); ?></h4>
     <form id="dateSelector" method="post">
@@ -175,4 +172,3 @@ $lastDays = $this->data['lastDays'];
         <div id="spsChart" class="pieChart"></div>
     </div>
 </div>
-</body>